Show option to link to DMOJ organization.
If the current course ID being viewed is not found in the dmoj_organize table yet, the option to "Link to DMOJ organization" (Yes/No) is displayed with the "OK" submit button (clicking on it won't do anything yet). If the current course ID is already found, the option "Link to DMOJ organization" is locked to "Yes" and cannot be changed, and the current DMOJ organization ID linked is displayed.