Add function to create DMOJ organization
Added function to create an organization on our DMOJ website using a POST request to {{domain}}/api/v2/organizations as shown in Postman. The function works as expected for testing purposes, but is not currently in the correct location (reloading the "Link to DMOJ organization" view.php page automatically creates a new organization, for testing purposes only).