Link between Moodle course and DMOJ organization
Partially completed the task "Create a link between between moodle account of member of a course and their DMOJ account" (https://docs.google.com/spreadsheets/d/1DEom0TfkXcMn_0pLvkwKuA-7QU7unEJ09R0y5-qfcss/edit?gid=1925042841#gid=1925042841). What's done so far: When the teacher of the course clicks "Link to DMOJ organization" and then chooses "Yes" to link, a new organization is created on DMOJ website via an API request with the same name as the course name, and participants on that Moodle course that have a linked DMOJ account is automatically added to the organization. If the Moodle participant doesn't have a linked DMOJ account yet, a DMOJ account is forced-created for them and linked to the Moodle participant account, then that new DMOJ account is added to the linked organization.

This is only done at the moment the Moodle teacher chooses to create a new organization. Next steps:
- Don't allow students to modify the state of "Link to DMOJ organization" (automatically finished because students cannot see "More" section in menu)
- Add a course boolean variable to determine whether the course is already linked to a DMOJ organization or not.
- Displays whether the course is linked to a DMOJ organization (and organization ID of it if already linked) in the course main page.
- Modify the "admin" list correctly (currently, only the "members" list is modified correctly)
- When a new participant is added to the Moodle course linked to a DMOJ organization, that participant is automatically added to DMOJ organization as well, using the same process described above when a new organization is linked.
- When an existing participant is removed from a linked Moodle course, that participant is automatically removed from the DMOJ organization as well (but not deleting the whole account).