Fully complete DMOJ organization linking
I couldn't find a way to do it automatically, so it has to be done manually by an editing teacher instead. DMOJ organization info will not be updated automatically when the Moodle course participant list changed or if members become editing teachers. Instead, it can be updated by having an editing teacher clicking on the "Force sync Moodle participants list to DMOJ organization member list" button.